Update auth libraries, separate "set" operations
The oauth2client library is deprecated. This update changes the sample to use the google-auth and google-auth-oauthlib libraries instead.

In addition, the code for setting video localizations has been updated so that if the default language is set, the code checks to see whether the video has a title and description in that language. If so, the script sets the video's title and description to the current title and description in the default language. (Without that step, the title and description for the new default language are changed to match the title and description of the old default language.)