Update auth libraries, add title/description args
The oauth2client library is deprecated. This update changes the sample to use the google-auth and google-auth-oauthlib libraries instead.

The changes also add --title and --description arguments for the playlist title and description.

A final note is that this script seems to be misnamed. This change did not change the functionality of the script, but it's likely a future change will either make this script update a playlist rather than insert one or provide an option for selecting an operation to perform for a playlist.