Update auth libraries, add options, documentation
The oauth2client library is deprecated. This update changes the sample to use the google-auth and google-auth-oauthlib libraries instead.

The updated script has improved options, allowing the user to set a content owner ID, job ID, report URL, and/or local filename using command-line arguments. The logic also now bypasses certain functions depending on the command-line arguments set. For example, if you specify the URL of the report that you want to download, the script doesn't prompt you to select a job and a report (since you've already done that).

The prerequisites section in the README file explains how to run this sample (and others in this directory).